src/server/services/pydantic-ai-agent/requirements.txt
pydantic-ai==0.0.19
pydantic-ai-slim[openai,anthropic]==0.0.19
ag-ui==0.0.35
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.0
httpx==0.27.0
python-multipart==0.0.9